# src/express-server/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: inf25dw1g05@gmail.com
    name: inf25dw1g05
  description: Torneio da Champions League
  title: Champions League
  version: 0.0.1
servers:
- description: Champions League
  url: http://localhost:8080
  variables:
    basePath:
      default: /
paths:
  /teams:
    get:
      description: "Ver todas as equipas, sendo possível filtrar!"
      operationId: retrieveTeams
      parameters:
      - description: Filtrar por país da equipa
        example: Portugal
        explode: true
        in: query
        name: country
        required: false
        schema:
          type: string
        style: form
      - description: Filtrar por grupo da equipa
        example: A
        explode: true
        in: query
        name: group
        required: false
        schema:
          type: string
        style: form
      - description: Filtrar por nome do treinador
        example: Ancelotti
        explode: true
        in: query
        name: coach
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Team"
                type: array
                xml:
                  name: Teams
          description: Array of Teams model instances
      summary: Ver equipas
      tags:
      - TeamsController
      x-eov-operation-handler: controllers/TeamsControllerController
    post:
      operationId: createTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Team"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
          description: Create a Team model instance
      summary: Criar equipa
      tags:
      - TeamsController
      x-eov-operation-handler: controllers/TeamsControllerController
...
components:
...
  schemas:
    Team:
      additionalProperties: false
      example:
        country: country
        name: name
        team_id: 0
        coach: coach
        group: group
      properties:
        team_id:
          format: int64
          type: integer
        name:
          type: string
        country:
          type: string
        coach:
          type: string
        group:
          type: string
      required:
      - coach
      - country
      - group
      - name
      title: Team
      type: object
...